+++ ncollide_geometry/partitioning/mod.rs
@@ -1,7 +1,6 @@
 //! Spatial partitioning tools.
 
-pub use partitioning::dbvt2::{DBVT2, DBVTLeaf2, DBVTLeafId};
-pub use partitioning::dbvt::{DBVTLeaf, DBVT};
+pub use partitioning::dbvt::{DBVT, DBVTLeaf, DBVTLeafId};
 pub use partitioning::bvt::{BVTNode, BinaryPartition, BVT};
 #[doc(inline)]
 pub use partitioning::bvt_visitor::{BVTVisitor, BoundingVolumeInterferencesCollector};
@@ -11,7 +10,6 @@ pub use partitioning::bvtt_visitor::BVTTVisitor;
 pub use partitioning::bvt_cost_fn::BVTCostFn;
 
 mod dbvt;
-mod dbvt2;
 mod bvt;
 
 #[doc(hidden)]
