+++ ncollide_geometry/query/contacts_internal/support_map_against_support_map.rs
@@ -6,12 +6,9 @@ use shape::{self, AnnotatedPoint, SupportMap};
 use query::algorithms::gjk::GJKResult;
 use query::algorithms::gjk;
 use query::algorithms::minkowski_sampling;
-use query::algorithms::epa3::{self, EPA3};
-use query::algorithms::epa2::{self, EPA2};
-use query::algorithms::simplex::Simplex;
-use query::algorithms::johnson_simplex::JohnsonSimplex;
-use query::algorithms::voronoi_simplex2::VoronoiSimplex2;
-use query::algorithms::voronoi_simplex3::VoronoiSimplex3;
+use query::algorithms::epa3;
+use query::algorithms::epa2;
+use query::algorithms::{EPA2, EPA3, Simplex, JohnsonSimplex, VoronoiSimplex2, VoronoiSimplex3};
 use query::Contact;
 use math::{Isometry, Point};
