+++ ncollide_geometry/query/point_internal/point_support_map.rs
@@ -3,10 +3,7 @@ use na;
 
 use query::algorithms::gjk;
 use query::algorithms::minkowski_sampling;
-use query::algorithms::simplex::Simplex;
-use query::algorithms::johnson_simplex::JohnsonSimplex;
-use query::algorithms::voronoi_simplex2::VoronoiSimplex2;
-use query::algorithms::voronoi_simplex3::VoronoiSimplex3;
+use query::algorithms::{Simplex, JohnsonSimplex, VoronoiSimplex2, VoronoiSimplex3};
 use query::{PointProjection, PointQuery};
 use shape::{Capsule, Cone, ConvexHull, Cylinder, SupportMap};
 use math::{Isometry, Point};
