+++ ncollide_geometry/query/ray_internal/ray_support_map.rs
@@ -4,10 +4,7 @@ use alga::general::Id;
 use na;
 
 use query::algorithms::gjk;
-use query::algorithms::simplex::Simplex;
-use query::algorithms::johnson_simplex::JohnsonSimplex;
-use query::algorithms::voronoi_simplex2::VoronoiSimplex2;
-use query::algorithms::voronoi_simplex3::VoronoiSimplex3;
+use query::algorithms::{Simplex, JohnsonSimplex, VoronoiSimplex2, VoronoiSimplex3};
 use query::{Ray, RayCast, RayIntersection};
 use shape::{Capsule, Cone, ConvexHull, Cylinder, MinkowskiSum, Segment, SupportMap};
 use math::{Isometry, Point};
