+++ ncollide_math/lib.rs
@@ -1,5 +1,14 @@
 //! Trait implemented by the primitive algebraic types used by ncollide.
 
+#![deny(non_camel_case_types)]
+#![deny(unused_parens)]
+#![deny(non_upper_case_globals)]
+#![deny(unused_qualifications)]
+#![deny(missing_docs)]
+#![deny(unused_results)]
+#![warn(unused_imports)]
+#![doc(html_root_url = "http://ncollide.org/rustdoc")]
+
 extern crate alga;
 extern crate approx;
 extern crate nalgebra as na;
