+++ ncollide_pipeline/narrow_phase/contact_generator/incremental_contact_manifold_generator.rs
@@ -55,6 +55,11 @@ where
             _matrix: PhantomData,
         }
     }
+
+    pub fn clear(&mut self) {
+        self.contacts.clear();
+        self.collector.clear();
+    }
 }
 
 impl<P, M, CD> IncrementalContactManifoldGenerator<P, M, CD>
